- Fixed a memory leaks in the zArray implementation. - Fixed minor warning. - Initial usage of zGetOpt.

+++ src/zConsole.cpp
@@ -11,7 +11,7 @@ zConsole::~zConsole(void) {}
 
 void zConsole::setColor(FgColor fg) {
 #if defined(WIN32)
-  SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), (int)(fg) | FOREGROUND_INTENSITY);
+  SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), (WORD)(fg) | FOREGROUND_INTENSITY);
 #else
   printf("%c[%d;%d;%dm", 0x1B, BRIGHT, (int)fg, 4);
 #endif
